[frontend] Addressed PR feedback
