+++ opencti-platform/opencti-front/src/private/components/locations/regions/RegionDeletion.tsx
@@ -1,57 +1,47 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
 import { graphql } from 'react-relay';
-import makeStyles from '@mui/styles/makeStyles';
 import { useNavigate } from 'react-router-dom';
 import { useFormatter } from '../../../../components/i18n';
 import Security from '../../../../utils/Security';
-import Transition from '../../../../components/Transition';
 import { KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
-import useDeletion from '../../../../utils/hooks/useDeletion';
 import useApiMutation from '../../../../utils/hooks/useApiMutation';
+import Transition from '../../../../components/Transition';
 
-// Deprecated - https://mui.com/system/styles/basics/
-// Do not use it for new code.
-const useStyles = makeStyles(() => ({
-  container: {
-    margin: 0,
-  },
-}));
-
-const CityPopoverDeletionDeleteMutation = graphql`
-  mutation CityPopoverDeletionDeleteMutation($id: ID!) {
-    cityEdit(id: $id) {
+const RegionDeletionDeleteMutation = graphql`
+  mutation RegionDeletionDeleteMutation($id: ID!) {
+    regionEdit(id: $id) {
       delete
     }
   }
 `;
 
-const CityPopoverDeletion = ({ id }: { id: string }) => {
-  const classes = useStyles();
+const RegionDeletion = ({ id }: { id: string }) => {
   const { t_i18n } = useFormatter();
   const navigate = useNavigate();
+  const [deleting, setDeleting] = useState<boolean>(false);
+  const [displayDelete, setDisplayDelete] = useState<boolean>(false);
   const deleteSuccessMessage = t_i18n('', {
     id: '... successfully deleted',
-    values: { entity_type: t_i18n('entity_City') },
+    values: { entity_type: t_i18n('entity_Region') },
   });
   const [commit] = useApiMutation(
-    CityPopoverDeletionDeleteMutation,
+    RegionDeletionDeleteMutation,
     undefined,
     { successMessage: deleteSuccessMessage },
   );
-  const handleClose = () => { };
-  const {
-    deleting,
-    handleOpenDelete,
-    displayDelete,
-    handleCloseDelete,
-    setDeleting,
-  } = useDeletion({ handleClose });
-
+  const handleClose = () => {};
+  const handleOpenDelete = () => {
+    setDisplayDelete(true);
+    handleClose();
+  };
+  const handleCloseDelete = () => {
+    setDisplayDelete(false);
+  };
   const submitDelete = () => {
     setDeleting(true);
     commit({
@@ -61,12 +51,13 @@ const CityPopoverDeletion = ({ id }: { id: string }) => {
       onCompleted: () => {
         setDeleting(false);
         handleClose();
-        navigate('/dashboard/locations/cities');
+        navigate('/dashboard/locations/regions');
       },
     });
   };
+
   return (
-    <div className={classes.container}>
+    <div style={{ margin: 0 }}>
       <Security needs={[KNOWLEDGE_KNUPDATE_KNDELETE]}>
         <Button
           color="error"
@@ -87,7 +78,7 @@ const CityPopoverDeletion = ({ id }: { id: string }) => {
       >
         <DialogContent>
           <DialogContentText>
-            {t_i18n('Do you want to delete this city?')}
+            {t_i18n('Do you want to delete this region?')}
           </DialogContentText>
         </DialogContent>
         <DialogActions>
@@ -103,4 +94,4 @@ const CityPopoverDeletion = ({ id }: { id: string }) => {
   );
 };
 
-export default CityPopoverDeletion;
+export default RegionDeletion;
